modifications to finetune locs test

+++ scripts/trainer_finetune.py
@@ -10,14 +10,22 @@ from safetensors.torch import load_file
 from data_utils.dataloader import prepare_dataloader
 
 from trainers.train_functions import (
-    commit_state, create_or_restore_data_state, create_or_restore_training_state_wandb, epoch_end_logs
+    create_or_restore_data_state, epoch_end_logs
 )
 from trainers.finetune_functions import (
-    finetune, unfreeze_base_model
+    finetune, create_training_state_finetune, init_wandb
 )
+
+from trainers.test_functions import (
+    get_class_probs
+)
+
+from sklearn.metrics import accuracy_score, roc_auc_score, average_precision_score
+
 from trainers import logger
 
 from accelerate import Accelerator
+from accelerate.utils import set_seed
 import json
 
 import argparse
@@ -51,7 +59,6 @@ if __name__ == "__main__":
     parser.add_argument("--base-model-path", type=str, required=True, help="Path to the base model state dictionary")
     parser.add_argument("--train-input", type=str, required=True, help="Path to train .npy file (samples, taxa, 2)")
     parser.add_argument("--best-dir", type=str, required=True, help="Directory to save best model so far")
-    parser.add_argument("--checkpoint-dir", type=str, required=True, help="Directory to save checkpoints for preemption")
     parser.add_argument("--data-restore-dir", type=str, required=True, help="Directory to restore data state")
     parser.add_argument("--vocab-restore-dir", type=str, required=True, help="Directory to restore vocab state")
     parser.add_argument("--batch-restore-dir", type=str, required=True, help="Directory to restore batch vocab state")
@@ -88,7 +95,6 @@ if __name__ == "__main__":
     base_model_path = args.base_model_path
     train_input = args.train_input
     best_dir = args.best_dir
-    checkpoint_dir = args.checkpoint_dir
     data_restore_dir = args.data_restore_dir
     vocab_restore_dir = args.vocab_restore_dir
     batch_restore_dir = args.batch_restore_dir
@@ -113,17 +119,17 @@ if __name__ == "__main__":
     model_config_path = args.model_config_path
     
     start_over = args.start_over
+    
+    accelerator = Accelerator(gradient_accumulation_steps=grad_accumulation_steps, mixed_precision="fp16" if enable_fp16 else "no", log_with="wandb" if wandb_enabled else None)
         
     # Set random seed for reproducibility
-    torch.manual_seed(args.seed)
-    np.random.seed(args.seed)
+    set_seed(args.seed, device_specific=True)
     
     with open(base_model_config_path, "r") as f:
         base_model_config = json.load(f)
         
     num_bins = base_model_config["n_input_bins"]
 
-    accelerator = Accelerator(gradient_accumulation_steps=grad_accumulation_steps, mixed_precision="fp16" if enable_fp16 else "no", log_with="wandb" if wandb_enabled else None)
 
     # Save notes to a markdown file in the save directory
     if accelerator.is_main_process and args.notes:
@@ -132,22 +138,24 @@ if __name__ == "__main__":
         with open(notes_path, "w") as notes_file:
             notes_file.write(args.notes)
 
-    if start_over:
+    if start_over and accelerator.is_main_process:
         logger.info("Starting over from scratch, deleting existing training state.")
         if os.path.exists(data_restore_dir):
+            # The above code is using the `shutil.rmtree()` function in Python to recursively remove a
+            # directory and all its contents. In this case, it is removing the directory specified by
+            # the variable `data_restore_dir`.
             shutil.rmtree(data_restore_dir)
         os.makedirs(data_restore_dir, exist_ok=True)
         if os.path.exists(batch_restore_dir):
             shutil.rmtree(batch_restore_dir)
         os.makedirs(batch_restore_dir, exist_ok=True)
         # don't remove vocab_restore_dir, as it is from pretraining
-        if os.path.exists(checkpoint_dir):
-            shutil.rmtree(checkpoint_dir)
-        os.makedirs(checkpoint_dir, exist_ok=True)
+    
+    accelerator.wait_for_everyone()
 
     # Create or restore data state
     train_data_dict, valid_data_dict, vocab, batch_vocab = create_or_restore_data_state(
-        train_input, num_bins, data_restore_dir, vocab_restore_dir, batch_restore_dir, accelerator, taxa_path=None, use_batch_labels=True, direct_batch_path=train_loc_labels_path, nrows=nrows
+        train_input, num_bins, data_restore_dir, vocab_restore_dir, batch_restore_dir, accelerator, taxa_path=None, use_batch_labels=True, direct_batch_path=train_loc_labels_path, nrows=nrows, seed=args.seed
     )
     
     check_vocab_basemodel_match(base_model_config, vocab)
@@ -185,6 +193,8 @@ if __name__ == "__main__":
         gen_percent=0,
         contrastive_embedding=False,
     )
+    
+    trainloader_len = len(train_loader)
 
     base_state_dict = load_file(base_model_path)
     
@@ -199,28 +209,33 @@ if __name__ == "__main__":
         with open(model_config_path, "w") as f:
             json.dump(model_config, f, indent=4)
 
-    total_steps = (frozen_max_epochs + unfrozen_max_epochs) * len(train_loader)
-    model, optimizer, scheduler, epoch, best_val_loss, patience_counter, extra_state = create_or_restore_training_state_wandb(
+    model, optimizer, scheduler = create_training_state_finetune(
         model_config,
         init_lr,
         cosine_warmup_ratio_or_step,
-        total_steps,
-        checkpoint_dir,
+        frozen_max_epochs * trainloader_len,
+        trainable_base_model=False,  # Set to False to freeze base model initially
+        base_state_dict=base_state_dict,
+    )
+    
+    init_wandb(
         wandb_enabled,
         wandb_entity,
         wandb_project,
         wandb_config,
-        is_pretrain=False,
-        accelerator=accelerator,
+        accelerator,
         wandb_run_name=wandb_run_name,
         wandb_run_notes=args.notes,
-        base_state_dict=base_state_dict,
-        trainable_base_model=False  # Set to False to freeze base model initially
     )
+    
+    # torch.save(model.state_dict(), "testdir/model_weights1.pth")
 
     train_loader, valid_loader, model, optimizer, scheduler = accelerator.prepare(
         train_loader, valid_loader, model, optimizer, scheduler
     )
+    epoch = 0
+    best_val_loss = float("inf")
+    patience_counter = 0
 
     while epoch < frozen_max_epochs:
         logger.info(f"Epoch {epoch + 1}/{frozen_max_epochs + unfrozen_max_epochs}")
@@ -256,14 +271,21 @@ if __name__ == "__main__":
             break
 
         epoch += 1
-        if accelerator.is_main_process:
-            commit_state(extra_state, epoch, best_val_loss, patience_counter, checkpoint_dir, accelerator)
-            
         accelerator.wait_for_everyone()
         
-    # now, train with base model unfrozen
-    unwrapped_model = accelerator.unwrap_model(model)
-    unfreeze_base_model(unwrapped_model, optimizer)
+    # now, train with base model unfrozen   
+    new_state_dict = load_file(os.path.join(best_dir, "model.safetensors"))
+    new_model, new_optimizer, new_scheduler = create_training_state_finetune(
+        model_config,
+        init_lr,
+        cosine_warmup_ratio_or_step,
+        unfrozen_max_epochs * trainloader_len,
+        trainable_base_model=True,  # Set to False to freeze base model initially
+        state_dict=new_state_dict,
+    )
+    new_model, new_optimizer, new_scheduler = accelerator.prepare(
+        new_model, new_optimizer, new_scheduler
+    )
     logger.info("Starting finetuning with base model unfrozen...")
     while epoch < frozen_max_epochs + unfrozen_max_epochs:
         logger.info(f"Epoch {epoch + 1}/{frozen_max_epochs + unfrozen_max_epochs}")
@@ -271,15 +293,15 @@ if __name__ == "__main__":
 
         # Train the model
         val_loss, val_acc = finetune(
-            model=model,
+            model=new_model,
             train_loader=train_loader,
             valid_loader=valid_loader,
             epoch=epoch,
             log_interval=log_interval,
             vocab=vocab,
             accelerator=accelerator,
-            optimizer=optimizer,
-            scheduler=scheduler,
+            optimizer=new_optimizer,
+            scheduler=new_scheduler,
             best_dir=best_dir,
             best_val_loss=best_val_loss,
             loss_fn=loss_fn,
@@ -299,10 +321,49 @@ if __name__ == "__main__":
             break
 
         epoch += 1
-        if accelerator.is_main_process:
-            commit_state(extra_state, epoch, best_val_loss, patience_counter, checkpoint_dir, accelerator)
 
-    logger.info("Training complete with best validation loss: {:.4f}".format(best_val_loss))
+    logger.info("Training complete with best validation loss: {:.4f}".format(best_val_loss))    
+    
+    # evaluate on eval set... test differences
+    probs, targets = get_class_probs(new_model, valid_loader, vocab.pad_index, accelerator)
+    probs = probs.cpu().numpy()
+    targets = targets.cpu().numpy()
+    
+    if accelerator.is_main_process:
+        print("shape of probs and targets is:", probs.shape, targets.shape)
+        print("location of probs and targets is", probs.device, targets.device)
+        predictions = np.argmax(probs, axis=1)
+        total_accuracy = accuracy_score(targets, predictions)
+        region_scores = []
+        for region, index in batch_vocab.stoi.items():
+            if region == "unknown":
+                continue
+            scores = probs[:, index]
+            binary_predictions = np.array((predictions == index), dtype=int)
+            binary_targets = np.array((targets == index), dtype=int)
+            accuracy = accuracy_score(binary_targets, binary_predictions)
+            auroc = roc_auc_score(binary_targets, scores)
+            aupr = average_precision_score(binary_targets, scores)
+            baseline_precision = np.mean(binary_targets)
+            
+            region_scores.append({
+                "Region": region,
+                "Accuracy": accuracy,
+                "AUC (ROC)": auroc,
+                "Average Precision": aupr,
+                "Baseline Precision": baseline_precision
+            })
+        
+        # Save the scores to a file
+        region_scores.sort(key=lambda x: x["Region"])
+        region_scores.append({"Total Accuracy": total_accuracy})
+        with open(os.path.join(args.best_dir, "valid_results.json"), "w") as f:
+            json.dump(region_scores, f, indent=4)
+
+        print(f"Scores for all regions saved to {args.best_dir}")
+        
     accelerator.end_training()
+    
+
